fix: logo dark mode color + other theme changes
