+++ src/assets/mainlogo/logo.module.css
@@ -3,8 +3,21 @@
     stroke-dashoffset: 1000;
     animation: dash 5s forwards normal ease-in-out;
     animation-iteration-count: 1;
+    fill: white;
+    stroke: #020547;
+    stroke-width: 6;
+    fill-opacity: 1;
 }
-  
+
+@media (prefers-color-scheme: dark) {
+    .path {
+        fill: #020547;
+        stroke: #EAB308;
+    }
+}
+
+
+
 @keyframes dash {
     from {
         stroke-dashoffset: 1050;
