+++ src/assets/mainlogo/main-logo.tsx
@@ -1,0 +1,55 @@
+import { SVGProps } from "react";
+import styles from './logo.module.css'
+
+const MainLogo = (props: SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    xmlnsXlink="http://www.w3.org/1999/xlink"
+    width={150}
+    zoomAndPan="magnify"
+    viewBox="0 0 375 135.000003"
+    height={54}
+    preserveAspectRatio="xMidYMid meet"
+    {...props}
+  >
+    <defs>
+      <g />
+    </defs>
+    <g>
+      <g transform="translate(27.482187, 118.745409)">
+        <g>
+          <path className={styles.path} d="M 65.507812 -93.246094 L 36.761719 -93.246094 C 31.304688 -93.132812 28.183594 -88.789062 28.183594 -84.667969 C 28.183594 -80.992188 30.636719 -77.539062 36.316406 -77.539062 L 48.015625 -77.539062 L 30.414062 -25.847656 C 27.40625 -17.15625 22.949219 -14.371094 17.714844 -14.371094 C 15.597656 -14.371094 14.035156 -16.710938 14.035156 -17.714844 C 14.148438 -18.160156 14.148438 -18.714844 14.148438 -19.160156 C 14.148438 -23.507812 10.027344 -26.179688 5.90625 -26.179688 C 1.78125 -26.179688 -2.449219 -23.507812 -2.449219 -17.046875 C -2.449219 -6.796875 7.128906 1.335938 19.382812 1.335938 C 28.742188 1.335938 41.21875 -3.898438 47.125 -20.945312 L 70.851562 -90.015625 C 71.410156 -91.574219 70.40625 -93.246094 68.625 -93.246094 Z M 65.507812 -93.246094 " />
+        </g>
+      </g>
+    </g>
+    <g>
+      <g transform="translate(84.849915, 118.745409)">
+        <g>
+          <path className={styles.path} d="M 69.851562 -37.65625 C 66.730469 -37.65625 63.277344 -35.761719 61.941406 -31.527344 C 57.597656 -17.824219 50.578125 -13.257812 45.566406 -13.257812 C 43.671875 -13.257812 42.554688 -14.261719 42.554688 -16.042969 C 42.554688 -16.933594 42.78125 -18.046875 43.113281 -19.273438 L 55.589844 -55.367188 C 56.035156 -56.703125 56.257812 -57.929688 56.257812 -59.042969 C 56.257812 -63.945312 52.25 -66.507812 48.125 -66.507812 C 45.007812 -66.507812 41.777344 -65.058594 39.996094 -62.273438 C 37.65625 -63.722656 33.308594 -65.394531 27.628906 -65.394531 C 20.277344 -65.394531 12.253906 -61.382812 9.246094 -52.804688 L -1.226562 -22.839844 C -1.894531 -20.832031 -2.339844 -18.828125 -2.339844 -16.820312 C -2.339844 -7.574219 5.125 0 14.929688 0 C 20.164062 0 25.511719 -2.339844 29.632812 -6.015625 C 33.199219 -0.890625 41.109375 0 44.449219 0 C 53.695312 0 66.0625 -4.679688 73.191406 -22.726562 C 74.753906 -27.070312 75.976562 -29.855469 75.976562 -31.863281 C 75.976562 -35.648438 73.082031 -37.65625 69.851562 -37.65625 Z M 33.085938 -38.769531 L 27.738281 -22.167969 C 25.398438 -14.929688 21.167969 -13.257812 18.828125 -13.257812 C 16.375 -13.257812 15.152344 -14.816406 15.152344 -17.378906 C 15.152344 -18.828125 15.484375 -20.164062 15.820312 -21.167969 L 22.949219 -43.558594 C 25.066406 -50.355469 28.183594 -51.246094 29.96875 -51.246094 C 32.085938 -51.246094 34.425781 -49.128906 34.425781 -45.339844 C 34.425781 -43.671875 33.976562 -41.554688 33.085938 -38.769531 Z M 33.085938 -38.769531 " />
+        </g>
+      </g>
+    </g>
+    <g>
+      <g transform="translate(143.331588, 118.745409)">
+        <g>
+          <path className={styles.path} d="M 98.703125 -37.320312 C 95.695312 -37.320312 92.464844 -35.648438 91.351562 -32.085938 C 86.449219 -18.714844 80.324219 -13.257812 75.085938 -13.257812 C 72.859375 -13.257812 71.746094 -14.59375 71.746094 -16.265625 C 71.746094 -16.820312 71.746094 -17.378906 71.96875 -17.824219 L 81.4375 -43.78125 C 82.105469 -45.898438 82.550781 -48.125 82.550781 -50.019531 C 82.550781 -59.488281 74.195312 -65.394531 64.058594 -65.394531 C 60.046875 -65.394531 54.253906 -64.390625 49.128906 -60.269531 C 46.789062 -62.945312 43.113281 -65.394531 36.652344 -65.394531 C 32.085938 -65.394531 29.410156 -64.058594 27.40625 -62.832031 C 26.179688 -64.503906 23.730469 -65.394531 21.277344 -65.394531 C 17.824219 -65.394531 14.261719 -63.722656 12.921875 -60.492188 L -4.902344 -9.136719 C -5.234375 -8.132812 -5.347656 -7.242188 -5.347656 -6.460938 C -5.347656 -2.339844 -1.78125 0 2.117188 0 C 5.90625 0 9.691406 -2.226562 11.25 -6.683594 L 24.507812 -44.785156 C 26.847656 -49.574219 30.414062 -52.25 33.753906 -52.25 C 35.984375 -52.25 37.542969 -51.023438 37.542969 -49.019531 C 37.542969 -48.460938 37.433594 -47.679688 37.097656 -46.902344 L 24.0625 -10.695312 C 23.507812 -9.246094 23.171875 -7.796875 23.171875 -6.574219 C 23.171875 -2.339844 26.515625 0 30.300781 0 C 33.976562 0 37.988281 -2.226562 39.660156 -7.019531 L 52.472656 -43.226562 C 54.476562 -48.796875 58.710938 -52.25 62.386719 -52.25 C 64.390625 -52.25 65.953125 -50.910156 65.953125 -48.796875 C 65.953125 -48.125 65.839844 -47.347656 65.507812 -46.34375 L 57.039062 -23.285156 C 56.148438 -20.722656 55.703125 -18.160156 55.703125 -15.820312 C 55.703125 -5.683594 64.058594 0 73.863281 0 C 84.890625 0 97.144531 -7.128906 104.832031 -29.410156 C 105.054688 -30.191406 105.277344 -30.96875 105.277344 -31.75 C 105.277344 -35.316406 102.046875 -37.320312 98.703125 -37.320312 Z M 98.703125 -37.320312 " />
+        </g>
+      </g>
+    </g>
+    <g>
+      <g transform="translate(230.887003, 118.745409)">
+        <g>
+          <path className={styles.path} d="M 47.125 -23.285156 C 47.125 -23.171875 47.125 -24.953125 46.679688 -27.070312 C 47.457031 -27.628906 48.238281 -28.519531 48.90625 -29.523438 C 50.800781 -32.308594 54.144531 -42 54.144531 -46.457031 C 54.144531 -57.820312 44.785156 -65.394531 32.753906 -65.394531 C 22.28125 -65.394531 12.367188 -59.601562 8.800781 -49.6875 L 0.667969 -26.738281 C -0.335938 -23.953125 -0.78125 -21.167969 -0.78125 -18.714844 C -0.78125 -7.019531 9.246094 0 20.609375 0 C 29.300781 0 37.210938 -2.785156 41.886719 -10.804688 C 43.894531 -10.25 46.234375 -9.914062 48.574219 -9.914062 C 61.71875 -9.914062 72.078125 -19.273438 75.867188 -30.96875 C 76.199219 -32.195312 76.421875 -33.308594 76.421875 -34.3125 C 76.421875 -38.433594 73.414062 -40.773438 70.183594 -40.773438 C 67.175781 -40.773438 64.058594 -38.992188 62.832031 -34.871094 C 60.382812 -27.960938 55.257812 -23.171875 48.683594 -23.171875 C 48.238281 -23.171875 47.570312 -23.171875 47.125 -23.285156 Z M 36.761719 -40.441406 C 36.429688 -39.4375 36.09375 -38.546875 35.984375 -37.65625 C 29.855469 -38.324219 24.398438 -33.308594 24.398438 -26.625 C 24.507812 -22.839844 26.402344 -20.832031 28.519531 -19.050781 C 27.070312 -15.152344 23.839844 -13.257812 20.386719 -13.257812 C 16.042969 -13.257812 14.929688 -15.929688 14.929688 -18.605469 C 14.929688 -20.722656 15.484375 -22.726562 15.929688 -23.839844 L 23.839844 -46.457031 C 25.175781 -50.132812 29.078125 -52.25 32.53125 -52.25 C 35.984375 -52.25 38.210938 -49.910156 38.210938 -47.347656 C 38.210938 -45.339844 37.542969 -42.78125 36.761719 -40.441406 Z M 36.761719 -40.441406 " />
+        </g>
+      </g>
+    </g>
+    <g>
+      <g transform="translate(289.034489, 118.745409)">
+        <g>
+          <path className={styles.path} d="M 62.164062 -32.085938 C 57.261719 -18.714844 51.136719 -13.257812 45.898438 -13.257812 C 43.894531 -13.257812 42.78125 -14.261719 42.78125 -16.152344 C 42.78125 -16.933594 43.003906 -17.824219 43.335938 -18.9375 L 52.027344 -43.671875 C 52.804688 -45.898438 53.25 -47.902344 53.25 -49.796875 C 53.25 -58.710938 45.675781 -65.394531 36.761719 -65.394531 C 32.417969 -65.394531 28.964844 -64.390625 26.070312 -62.273438 C 25.398438 -64.28125 22.613281 -65.394531 19.605469 -65.394531 C 16.488281 -65.394531 13.480469 -63.945312 12.367188 -60.828125 L -6.238281 -8.800781 C -6.574219 -7.796875 -6.683594 -6.90625 -6.683594 -6.015625 C -6.683594 -1.894531 -3.453125 0.222656 0.113281 0.222656 C 4.011719 0.222656 8.464844 -2.226562 10.136719 -7.242188 C 14.035156 -18.046875 19.382812 -33.308594 23.171875 -44.003906 C 26.179688 -49.574219 30.078125 -52.25 33.199219 -52.25 C 35.203125 -52.25 36.652344 -50.910156 36.652344 -48.796875 C 36.652344 -48.015625 36.429688 -47.011719 36.09375 -46.011719 L 27.960938 -23.0625 C 27.070312 -20.5 26.515625 -18.046875 26.515625 -15.707031 C 26.515625 -5.457031 34.980469 0 44.449219 0 C 55.8125 0 67.957031 -7.6875 75.644531 -29.410156 C 75.976562 -30.300781 76.199219 -31.191406 76.199219 -31.972656 C 76.199219 -35.539062 72.96875 -37.433594 69.628906 -37.433594 C 66.621094 -37.433594 63.277344 -35.761719 62.164062 -32.085938 Z M 62.164062 -32.085938 " />
+        </g>
+      </g>
+    </g>
+  </svg>
+);
+export default MainLogo;
